requirements.txt
Django==5.0.3

psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==24.1
django-environ==0.11.2
